Refactor: RepositoryInfoをmodelパッケージに移動
